feat(min-spend): Extract terminated dates service method into a separate service
