+++ app/services/invoices/calculate_fees_service.rb
@@ -19,7 +19,11 @@ module Invoices
       ActiveRecord::Base.transaction do
         invoice.invoice_subscriptions.each do |invoice_subscription|
           subscription = invoice_subscription.subscription
-          date_service = terminated_date_service(subscription, date_service(subscription))
+          date_service = Subscriptions::TerminatedDatesService.new(
+            subscription:,
+            invoice:,
+            date_service: date_service(subscription),
+          ).call
 
           boundaries = {
             from_datetime: invoice_subscription.from_datetime,
@@ -58,27 +62,6 @@ module Invoices
       result.record_validation_failure!(record: e.record)
     end
 
-    def terminated_date_service(subscription, date_service)
-      return date_service unless subscription.terminated? && subscription.next_subscription.nil?
-
-      # First we need to ensure that termination date is not started_at date. In that case boundaries are correct
-      # and we should bill only one day. If this is not the case we should proceed.
-      return date_service if (timestamp - 1.day) < subscription.started_at
-
-      # Date service has various checks for terminated subscriptions. We want to avoid it and fetch boundaries
-      # for current usage (current period) but when subscription was active (one day ago)
-      duplicate = subscription.dup.tap { |s| s.status = :active }
-      new_dates_service = Subscriptions::DatesService.new_instance(duplicate, timestamp - 1.day, current_usage: true)
-
-      return date_service if timestamp < new_dates_service.charges_to_datetime
-      return date_service unless (timestamp - new_dates_service.charges_to_datetime) < 1.day
-
-      # We should calculate boundaries as if subscription was not terminated
-      new_dates_service = Subscriptions::DatesService.new_instance(duplicate, timestamp, current_usage: false)
-
-      matching_invoice_subscription?(subscription, new_dates_service) ? date_service : new_dates_service
-    end
-
     private
 
     attr_accessor :invoice, :subscriptions, :timestamp, :recurring, :context
@@ -95,22 +78,6 @@ module Invoices
         timestamp,
         current_usage: subscription.terminated? && subscription.upgraded?,
       )
-    end
-
-    def matching_invoice_subscription?(subscription, date_service)
-      base_query = InvoiceSubscription
-        .where(subscription_id: subscription.id)
-        .recurring
-        .where(from_datetime: date_service.from_datetime)
-        .where(to_datetime: date_service.to_datetime)
-
-      if subscription.plan.yearly? && subscription.plan.bill_charges_monthly?
-        base_query = base_query
-          .where(charges_from_datetime: date_service.charges_from_datetime)
-          .where(charges_to_datetime: date_service.charges_to_datetime)
-      end
-
-      base_query.exists?
     end
 
     def create_minimum_commitment_true_up_fee(invoice_subscription)
